Move schedule generation to module
