+++ src/scheduling/__init__.py
@@ -1,2 +1,6 @@
 """Scheduling related types and functions."""
 from .types import *
+from .setup_lp import *
+
+from .simple_schedule import generate_simple_schedule
+from .extended_schedule import generate_extended_schedule
